requirements.txt
streamlit
pandas
numpy
faiss-cpu
sentence-transformers
langchain
gdown
requests
beautifulsoup4
torch
python-dotenv
